+++ packages/flame/src/InputGroup/InputGroup.tsx
@@ -24,7 +24,7 @@ const InputGroupAddon = styled(Flex)<InputGroupAddonProps>`
   )};
 `;
 
-const InputGroup: React.FC = ({ children, ...restProps }) => {
+const InputGroup: React.FC<FlameFlexProps> = ({ children, ...restProps }) => {
   const nextChildren = React.Children.map(children, (child: any, index) => {
     if (index === 0) {
       return React.cloneElement(child, {